Move DataSetChangeAPI to api module
