+++ hq_superset/__init__.py
@@ -19,8 +19,7 @@ def flask_app_mutator(app):
     appbuilder.add_view(views.HQDatasourceView, 'Update HQ Datasource', menu_cond=lambda *_: False)
     appbuilder.add_view(views.SelectDomainView, 'Select a Domain', menu_cond=lambda *_: False)
     appbuilder.add_api(api.OAuth)
-
-    appbuilder.add_view(views.DataSetChangeAPI, 'DataSet Change API', menu_cond=lambda *_: False)
+    appbuilder.add_api(api.DataSetChangeAPI)
     app.before_request_funcs.setdefault(None, []).append(
         hq_domain.before_request_hook
     )
